+++ actionpack/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Make `Session#merge!` stringify keys.
 
+    Previously `Session#update` would, but `merge!` wouldn't.
+
+    *Drew Bragg*
 
 Please check [7-0-stable](https://github.com/rails/rails/blob/7-0-stable/actionpack/CHANGELOG.md) for previous changes.